Switch to using XAudio2 for sound alert so volume can be set (through the "soundFileVolume" setting).
Add options to specify thresholds for wisdom/portal scrolls and scraps/whetstones.
When there is less of a currency than the threshold in the currency tab, add that currency to the loot filter.
The loot filter will not be automatically re-written just for this - use a force update in that case.
See README.md for details.